requirements.txt
accelerate
ctransformers
deepspeed==0.11.1+e9503fe
extensions
gradio
Jinja2
langchain
modules
numpy
Pillow
psutil
PyYAML
Requests
rich
safetensors
torch
tqdm
transformers
treelib